move ETH specific test from p2p to trinity module
